+++ packages/styleguide/.storybook/components/Box.tsx
@@ -1,0 +1,4 @@
+import { styled } from '@storybook/theming';
+import { allProps } from './styles';
+
+export const Box = styled.div(allProps);
