+++ packages/styleguide/.storybook/components/index.tsx
@@ -5,3 +5,5 @@ export * from './AboutHeading';
 export * from './PropsTable';
 export * from './DataTable';
 export * from './MarkdownElements';
+export * from './Box';
+export * from './ColorScale';
